// src/enums/funds-transfer-calculated-status.ts
export enum FundsTransferCalculatedStatus {
    PENDING = 'pending',
    NOT_PAID = 'not_paid',
    PAID = 'paid',
    PENDING_DOCUMENTS = 'pending_documents',
    DAMAGE_CLAIM = 'damage_claim',
    PENDING_APPROVAL = 'pending_approval',
    ACTION_REQUIRED = 'action_required',
    FACTORING_REQUESTED = 'factoring_requested',
    FACTORING_CONFIRMED = 'factoring_confirmed',
    READY_FOR_FACTORING = 'ready_for_factoring',
    PENDING_DELIVERY = 'pending_delivery',
    AWAITING_CONFIRMATION = 'awaiting_confirmation',
}